add PID validation to prevent overflow

+++ src/client.c
@@ -6,7 +6,7 @@
 /*   By: oprosvir <oprosvir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/27 23:01:49 by oprosvir          #+#    #+#             */
-/*   Updated: 2024/03/10 13:12:35 by oprosvir         ###   ########.fr       */
+/*   Updated: 2024/05/27 23:02:49 by oprosvir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,6 +51,11 @@ int	main(int argc, char *argv[])
 		exit(EXIT_FAILURE);
 	}
 	server_pid = ft_atoi(argv[1]);
+	if (server_pid <= 0 || server_pid > 99998)
+	{
+		ft_printf("ERROR: Invalid PID: %d\n", server_pid);
+		exit(EXIT_FAILURE);
+	}
 	if (kill(server_pid, 0) < 0)
 	{
 		ft_printf("ERROR: Failed to send signal to PID: %d\n", server_pid);
